+++ DataManager/ClientApp/minerva/ts/core/index.ts
@@ -1,6 +1,7 @@
 export * from './shared';
 export * from './http';
 export * from './user';
+export * from './products';
 
 export type NavItem = {
   navId: string,