+++ DataManager/ClientApp/minerva/ts/core/shared.ts
@@ -1,4 +1,4 @@
-import $ from 'jquery/dist/jquery.slim';
+import $ from 'jquery/slim';
 
 export const query = (selectors: string) => $(selectors);
 export const beforeProcessing = () => query('.loading-indicator').show();