+++ DataManager/ClientApp/minerva/ts/views/ads-manager.view.tsx
@@ -1,0 +1,26 @@
+export function AdsManagerView() {
+  // const [selectedCampaigns, setSelectedCampaigns] = useState<CampaignInfo[]>([]), // For bulk actions
+  //       [showModal, setShowModal] = useState(false);
+
+  // useEffect(() => {
+  //   CampaignInfoStore.loadWithSplashScreen();
+  // }, []);
+
+  // async function bulkGenerate() {
+  //   const campaigns = await CampaignGenerationModel.fetch();
+  //   afterProcessing();
+  //   console.log({campaigns});
+  //   CampaignInfoStore.loadData(campaigns);
+  // };
+
+  // function openNew() {}
+
+  // function openEdit(id: number) {}
+
+  // function toggleActive(id: number) {}
+
+  // function deleteCampaign(id: number) {}
+
+  return <>
+  </>
+}