+++ DataManager/ClientApp/minerva/ts/views/products/dashboard.view.tsx
@@ -1,0 +1,7 @@
+import { ProductLayout } from './product-layout.component';
+
+export function DashboardView() {
+  return <ProductLayout>
+    <div>Dashboard</div>
+  </ProductLayout>
+}