+++ DataManager/ClientApp/minerva/ts/views/products/index.ts
@@ -1,0 +1,3 @@
+export * from './list.view';
+export * from './settings.view';
+export * from './dashboard.view';