+++ DataManager/ClientApp/minerva/ts/views/products/settings.view.tsx
@@ -1,0 +1,7 @@
+import { ProductLayout } from './product-layout.component';
+
+export function ProductSettingsView() {
+  return <ProductLayout>
+      <div>Settings</div>
+    </ProductLayout>
+}