+++ DataManager/Shared/IDataService.cs
@@ -1,0 +1,11 @@
+namespace DataManager.Shared;
+
+public abstract class IDataService<T>(IDataStore<T> dataStore) {
+  public IEnumerable<T> List() => dataStore.List();
+}
+
+public interface IDataStore<T> {
+  IEnumerable<T> List();
+}
+
+public record NavItem(string NavId, string NavName, string NavIcon, string NavPath, List<NavItem> Children);