+++ models/shared/curated/user_profile.sql
@@ -41,7 +41,7 @@ with user_registration as (
   from {{
     source({
       'name': 'user_profile',
-      'location': '{{lakehouse.location}}/{{product_id}}/enriched/user_profile',
+      'location': '{{lakehouse.location}}/{{product_id}}/curated/user_profile',
     })
   }}
   union (